+++ app/scripts/filters/goodBadIssue.js
@@ -13,7 +13,7 @@ angular.module('wallboardApp')
             } else if (input['MINOR'] > 0 || input['INFO'] > 0) {
                 return 'neutral';
             } else if (input['BLOCKER'] == 0, input['CRITICAL'] == 0, input['MAJOR'] == 0, input['MINOR'] == 0, input['INFO'] == 0) {
-                return 'good';
+                return 'neutral';
             } else
                 return 'failure';
         };